requirements.txt
# Core data
xarray>=2024.1
netCDF4>=1.7
h5netcdf>=1.3
requests>=2.31
tqdm>=4.66

# Geospatial
geopandas>=1.0
shapely>=2.0
fiona>=1.10
pyproj>=3.6

# ML - Classical
scikit-learn>=1.5
xgboost>=2.0
torch>=2.5
torchvision>=0.20

# ML - Quantum
qiskit>=2.0
qiskit-machine-learning>=0.9
qiskit-aer>=0.15
qiskit-ibm-runtime>=0.34

# Visualization
matplotlib>=3.9
plotly>=5.24
kaleido>=0.2
seaborn>=0.13

# Utilities
pyyaml>=6.0
joblib>=1.4
pandas>=2.2
numpy>=1.26
scipy>=1.13
scikit-learn>=1.5
ipykernel>=6.29
notebook>=7.2
